+++ package.json
@@ -7,6 +7,7 @@
     "cors": "^2.8.5",
     "ejs": "^2.6.2",
     "node-sass": "^4.12.0",
+    "prop-types": "^15.7.2",
     "proxy-middleware": "^0.15.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
@@ -32,5 +33,13 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.17.3",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.14.0",
+    "prettier-eslint": "^9.0.0"
   }
 }
