--- src/App.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import TodoList from './conponent/TodoList';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <TodoList />
-    </div>
-  );
-}
-
-export default App;
