+++ src/App.test.js
@@ -1,6 +1,8 @@
+/* eslint-disable react/jsx-filename-extension */
+/* eslint-disable no-undef */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './index';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
