+++ src/conponent/TodoList/TodoList.scss
@@ -1,4 +1,4 @@
-#body{
+.body{
   width: 100%;
   text-align: center;
   height: 100%;
